+++ CHANGELOG.md
@@ -32,10 +32,13 @@
   what's embedded in the OSS telepresence client.
 
 - Bugfix: Telepresence traffic-manager extracts the cluster domain (e.g. "cluster.local") using a CNAME lookup for "kubernetes.default"
-  instead of "kubernetes.default.svc".
+  instead of "kubernetes.default.svc". [PR 2959](https://github.com/telepresenceio/telepresence/pull/2959).
 
 - Bugfix: A timeout was added to the pre-delete hook `uninstall-agents`, so that a helm uninstall doesn't
   hang when there is no running traffic-manager.
+
+- Bugfix: The `Helm.Revision` is now used to prevent that Helm hook calls are served by the wrong revision of
+  the traffic-manager. [Issue 2954](https://github.com/telepresenceio/telepresence/issues/2954).
 
 ### 2.9.5 (December 8, 2022)